Put parser in separate module
